Separated stream groupings with topology groupings.
