+++ src/main/java/grouping/stream/ZoneShuffleGrouping.java
@@ -1,4 +1,4 @@
-package grouping;
+package grouping.stream;
 
 import backtype.storm.generated.GlobalStreamId;
 import backtype.storm.task.WorkerTopologyContext;
